+++ packages/python_ffi/lib/src/python_ffi_base.dart
@@ -1,6 +1,9 @@
 part of python_ffi;
 
 class PythonFfi extends PythonFfiBase with PythonFfiMixin {
+  @visibleForTesting
+  factory PythonFfi() => PythonFfi._();
+
   PythonFfi._();
 
   static PythonFfi? _instance;
@@ -22,6 +25,13 @@ class PythonFfi extends PythonFfiBase with PythonFfiMixin {
   @override
   String get name => "PythonFfi";
 
-  FutureOr<void> initialize() async =>
-      await PythonFfiDelegate.instance.initialize();
+  /// Initializes the Python runtime and prepares bundled Python modules for
+  /// import.
+  ///
+  /// *Note:* Supply [package] if you are implementing a package that uses this
+  /// plugin. This will ensure that the Python runtime is initialized with the
+  /// correct path to the bundled Python modules when your package is used in a
+  /// Flutter app.
+  FutureOr<void> initialize({String? package}) async =>
+      await PythonFfiDelegate.instance.initialize(package: package);
 }