+++ packages/python_ffi_cpython/lib/python_ffi_cpython.dart
@@ -15,15 +15,27 @@ import "package:python_ffi_interface/python_ffi_interface.dart";
 /// The macOS and Windows implementation of [PythonFfiPlatform].
 final class PythonFfiCPython extends PythonFfiCPythonBase
     with PythonFfiCPythonMixin {
+  String? _package;
+
+  @override
+  Future<void> initialize({required String? package}) {
+    _package = package;
+    return super.initialize(package: package);
+  }
+
   @override
   Future<Directory> getApplicationSupportDirectory() async =>
       path_provider.getApplicationSupportDirectory();
 
   @override
   FutureOr<ByteData> loadPythonFile(PythonSourceFileEntity sourceFile) {
+    final String? package = _package;
+    final String assetPath = "${switch (package) {
+      String() => "packages/$package/",
+      _ => "",
+    }}python-modules/${sourceFile.name}";
     if (sourceFile is SourceFile) {
-      return PlatformAssetBundle()
-          .load("python-modules/${sourceFile.name}");
+      return PlatformAssetBundle().load(assetPath);
     } else if (sourceFile is SourceBase64) {
       return ByteData.view(base64Decode(sourceFile.base64).buffer);
     }
@@ -141,10 +153,17 @@ final class PythonFfiCPython extends PythonFfiCPythonBase
   }
 
   @override
-  Future<Set<PythonModuleDefinition>> discoverPythonModules() async {
+  Future<Set<PythonModuleDefinition>> discoverPythonModules({
+    required String? package,
+  }) async {
     try {
-      final ByteData modulesJsonRaw = await PlatformAssetBundle()
-          .load("python-modules/modules.json");
+      const String localModulesJsonPath = "python-modules/modules.json";
+      final String modulesJsonPath = switch (package) {
+        String() => "packages/$package/$localModulesJsonPath",
+        _ => localModulesJsonPath,
+      };
+      final ByteData modulesJsonRaw =
+          await PlatformAssetBundle().load(modulesJsonPath);
       return _decodePythonModules(
         utf8.decode(modulesJsonRaw.buffer.asUint8List()),
       ).toSet();